// enums/photos-fields.ts
export enum PhotosFieldsEnum {
  PRODUCT = '1. FOTO DEL PRODUCTO en el Estante',
  PROMOTOR = '2. FOTO DE LA PROMOTORA con la Mesa de Activación',
  CLIENTS = 'FOTO DE CLIENTES',
  OTHER = 'Foto',
}

export enum PhotoTypesEnum {
  PRODUCT = 'product',
  PROMOTOR = 'promotor',
  CLIENT = 'client',
  OTHER = 'other',
}